+++ packages/server/src/dal/sql/abstract-sql-db-engine.ts
@@ -1,4 +1,4 @@
-import {injectable} from "inversify";
+import {inject, injectable, interfaces} from "inversify";
 import AclEntryModel from "./models/acl-entry-model";
 import {
     ACL_ENTRY, AGE,
@@ -66,10 +66,78 @@ import AgeModel from "./models/calendar/age-model";
 import MonthModel from "./models/calendar/month-model";
 import DayOfTheWeekModel from "./models/calendar/day-of-the-week-model";
 import WikiPageToWikiPageModel from "./models/wiki-page-to-wiki-page-model";
+import {DatabaseContext} from "../database-context";
+import SQLDatabaseContext from "./sql-database-context";
+import {INJECTABLE_TYPES} from "../../di/injectable-types";
+import {ArticleRepository} from "../repository/article-repository";
+import {CalendarRepository} from "../repository/calendar-repository";
+import {ChunkRepository} from "../repository/chunk-repository";
+import EventWikiRepository from "../repository/event-wiki-repository";
+import {FileRepository} from "../repository/file-repository";
+import FogStrokeRepository from "../repository/fog-stroke-repository";
+import {GameRepository} from "../repository/game-repository";
+import {ImageRepository} from "../repository/image-repository";
+import {ItemRepository} from "../repository/item-repository";
+import {ModelRepository} from "../repository/model-repository";
+import {MonsterRepository} from "../repository/monster-repository";
+import {PersonRepository} from "../repository/person-repository";
+import {PinRepository} from "../repository/pin-repository";
+import {PlaceRepository} from "../repository/place-repository";
+import {RoleRepository} from "../repository/role-repository";
+import {ServerConfigRepository} from "../repository/server-config-repository";
+import StrokeRepository from "../repository/stroke-repository";
+import {UserRepository} from "../repository/user-repository";
+import {WikiFolderRepository} from "../repository/wiki-folder-repository";
+import {WikiPageRepository} from "../repository/wiki-page-repository";
+import {WorldRepository} from "../repository/world-repository";
 
 
 @injectable()
 export default class AbstractSqlDbEngine {
+
+
+    @inject(INJECTABLE_TYPES.ArticleRepository)
+    articleRepository: ArticleRepository;
+    @inject(INJECTABLE_TYPES.CalendarRepository)
+    calendarRepository: CalendarRepository;
+    @inject(INJECTABLE_TYPES.ChunkRepository)
+    chunkRepository: ChunkRepository;
+    @inject(INJECTABLE_TYPES.EventWikiRepository)
+    eventRepository: EventWikiRepository;
+    @inject(INJECTABLE_TYPES.FileRepository)
+    fileRepository: FileRepository;
+    @inject(INJECTABLE_TYPES.FogStrokeRepository)
+    fogStrokeRepository: FogStrokeRepository;
+    @inject(INJECTABLE_TYPES.GameRepository)
+    gameRepository: GameRepository;
+    @inject(INJECTABLE_TYPES.ImageRepository)
+    imageRepository: ImageRepository;
+    @inject(INJECTABLE_TYPES.ItemRepository)
+    itemRepository: ItemRepository;
+    @inject(INJECTABLE_TYPES.ModelRepository)
+    modelRepository: ModelRepository;
+    @inject(INJECTABLE_TYPES.MonsterRepository)
+    monsterRepository: MonsterRepository;
+    @inject(INJECTABLE_TYPES.PersonRepository)
+    personRepository: PersonRepository;
+    @inject(INJECTABLE_TYPES.PinRepository)
+    pinRepository: PinRepository;
+    @inject(INJECTABLE_TYPES.PlaceRepository)
+    placeRepository: PlaceRepository;
+    @inject(INJECTABLE_TYPES.RoleRepository)
+    roleRepository: RoleRepository;
+    @inject(INJECTABLE_TYPES.ServerConfigRepository)
+    serverConfigRepository: ServerConfigRepository;
+    @inject(INJECTABLE_TYPES.StrokeRepository)
+    strokeRepository: StrokeRepository;
+    @inject(INJECTABLE_TYPES.UserRepository)
+    userRepository: UserRepository;
+    @inject(INJECTABLE_TYPES.WikiFolderRepository)
+    wikiFolderRepository: WikiFolderRepository;
+    @inject(INJECTABLE_TYPES.WikiPageRepository)
+    wikiPageRepository: WikiPageRepository;
+    @inject(INJECTABLE_TYPES.WorldRepository)
+    worldRepository: WorldRepository;
     
     connectAll(connection: Sequelize) {
         // Is there any better way to do this? How to handle a bunch of static methods with the same signature?
@@ -175,6 +243,33 @@ export default class AbstractSqlDbEngine {
             logger: console,
         });
         await umzug.up();
+    }
+
+    async createDatabaseContext(connection: Sequelize): Promise<DatabaseContext> {
+        return new SQLDatabaseContext(
+            connection,
+            this.articleRepository,
+            this.calendarRepository,
+            this.chunkRepository,
+            this.eventRepository,
+            this.fileRepository,
+            this.fogStrokeRepository,
+            this.gameRepository,
+            this.imageRepository,
+            this.itemRepository,
+            this.modelRepository,
+            this.monsterRepository,
+            this.personRepository,
+            this.pinRepository,
+            this.placeRepository,
+            this.roleRepository,
+            this.serverConfigRepository,
+            this.strokeRepository,
+            this.userRepository,
+            this.wikiFolderRepository,
+            this.wikiPageRepository,
+            this.worldRepository,
+        );
     }
     
 }