+++ packages/server/src/seeders/role-seeder.ts
@@ -9,46 +9,39 @@ import RoleFactory from "../domain-entities/factory/role-factory";
 import {RoleRepository} from "../dal/repository/role-repository";
 import {ServerConfigRepository} from "../dal/repository/server-config-repository";
 import {UserRepository} from "../dal/repository/user-repository";
+import {DatabaseContext} from "../dal/database-context";
 
 @injectable()
 export class RoleSeeder implements Seeder {
-	@inject(INJECTABLE_TYPES.UserRepository)
-	userRepository: UserRepository;
-
-	@inject(INJECTABLE_TYPES.ServerConfigRepository)
-	serverConfigRepository: ServerConfigRepository;
-
-	@inject(INJECTABLE_TYPES.RoleRepository)
-	roleRepository: RoleRepository;
 
 	@inject(INJECTABLE_TYPES.RoleFactory)
 	roleFactory: RoleFactory;
 
-	seed = async (): Promise<void> => {
-		let allUsersRole: Role = await this.roleRepository.findOneByName(EVERYONE);
+	seed = async (databaseContext: DatabaseContext): Promise<void> => {
+		let allUsersRole: Role = await databaseContext.roleRepository.findOneByName(EVERYONE);
 		if (!allUsersRole) {
-			const server = await this.serverConfigRepository.findOne();
+			const server = await databaseContext.serverConfigRepository.findOne();
 			if (!server) {
 				throw new Error("Server needs to exist!");
 			}
 			allUsersRole = this.roleFactory.build({name: EVERYONE, world: null, acl: []});
-			await this.roleRepository.create(allUsersRole);
+			await databaseContext.roleRepository.create(allUsersRole);
 			console.log(`Created default role "${EVERYONE}"`);
 		}
-		let loggedInRole: Role = await this.roleRepository.findOneByName(LOGGED_IN);
+		let loggedInRole: Role = await databaseContext.roleRepository.findOneByName(LOGGED_IN);
 		if (!loggedInRole) {
-			const server = await this.serverConfigRepository.findOne();
+			const server = await databaseContext.serverConfigRepository.findOne();
 			if (!server) {
 				throw new Error("Server needs to exist!");
 			}
 			loggedInRole = this.roleFactory.build({name: LOGGED_IN, world: null, acl: []});
-			await this.roleRepository.create(loggedInRole);
+			await databaseContext.roleRepository.create(loggedInRole);
 			server.acl.push({
 				permission: WORLD_CREATE,
 				principal: loggedInRole._id,
 				principalType: ROLE
 			});
-			await this.serverConfigRepository.update(server);
+			await databaseContext.serverConfigRepository.update(server);
 			console.log(`Created default role "${LOGGED_IN}"`);
 		}
 	};