+++ packages/server/src/seeders/server-config-seeder.ts
@@ -3,19 +3,16 @@ import { v4 as uuidv4 } from "uuid";
 import { inject, injectable } from "inversify";
 import { INJECTABLE_TYPES } from "../di/injectable-types";
 import ServerConfigFactory from "../domain-entities/factory/server-config-factory";
-import {ServerConfigRepository} from "../dal/repository/server-config-repository";
+import {DatabaseContext} from "../dal/database-context";
 
 @injectable()
 export class ServerConfigSeeder implements Seeder {
 
-	@inject(INJECTABLE_TYPES.ServerConfigRepository)
-	serverConfigRepository: ServerConfigRepository;
-
 	@inject(INJECTABLE_TYPES.ServerConfigFactory)
 	serverConfigFactory: ServerConfigFactory;
 
-	seed = async (): Promise<void> => {
-		let server = await this.serverConfigRepository.findOne();
+	seed = async (databaseContext: DatabaseContext): Promise<void> => {
+		let server = await databaseContext.serverConfigRepository.findOne();
 		if (!server) {
 			const unlockCode: string = process.env.UNLOCK_CODE || uuidv4();
 			let registerCodes: string[] = [];
@@ -23,7 +20,7 @@ export class ServerConfigSeeder implements Seeder {
 				registerCodes = process.env.REGISTER_CODES.split(',');
 			}
 			server = this.serverConfigFactory.build({version: "1.0", registerCodes, adminUsers: [], unlockCode, acl: [], defaultWorld: null});
-			await this.serverConfigRepository.create(server);
+			await databaseContext.serverConfigRepository.create(server);
 		}
 	};
 }