+++ packages/server/tests/integration/setup.ts
@@ -5,12 +5,10 @@ import { container, bindAll } from "../../src/di/inversify";
 import { INJECTABLE_TYPES } from "../../src/di/injectable-types";
 import {DbEngine, SessionContextFactory} from "../../src/types";
 import {ServerConfigService} from "../../src/services/server-config-service";
-import {Factory} from "../../src/types";
 import {MockSessionContextFactory} from "./mock-session-context-factory";
 import {DefaultTestingContext} from "./default-testing-context";
 import {TEST_INJECTABLE_TYPES} from "./injectable-types";
 import RpgToolsServer from "../../src/server/rpgtools-server";
-import {DatabaseContext} from "../../src/dal/database-context";
 import {jest} from '@jest/globals'
 
 process.env.ACCESS_TOKEN_SECRET = "asdf1234";
@@ -36,14 +34,11 @@ beforeAll(async () => {
 	dbEngine.setDbHost("localhost");
 	await dbEngine.changeDb(process.env.TEST_SUITE);
 	await dbEngine.connect();
-	await server.seedDB();
+	const databaseContext = await dbEngine.createDatabaseContext();
+	await server.seedDB(databaseContext);
 	const service = container.get<ServerConfigService>(INJECTABLE_TYPES.ServerConfigService);
-	const databaseContextFactory = container.get<Factory<DatabaseContext>>(INJECTABLE_TYPES.DatabaseContextFactory);
-	const session = await dbEngine.createDatabaseSession();
-	const databaseContext = databaseContextFactory({session});
 	const serverConfig = await service.getServerConfig(databaseContext);
 	await service.unlockServer(serverConfig.unlockCode, "tester@gmail.com", "tester", "tester", databaseContext);
-	await session.commit();
 });
 
 afterAll(async function () {